Fix cs:wizard reauth.
